Move StarknetConfig to separate file
